Some bug fixes, polarity and trigger not working properly (for both Gretina and Gretina4), added TTCL type for Gretina4.
git-svn-id: svn://orca.physics.unc.edu/Orca@210 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Vme/Gretina4/ORGretina4Model.h
@@ -221,7 +221,7 @@ enum Gretina4FIFOStates {
 #pragma mark ¥¥¥Hardware Access
 - (short) readBoardID;
 - (void) initBoard;
-- (short) readControlReg:(int)channel;
+- (unsigned long) readControlReg:(int)channel;
 - (void) writeControlReg:(int)channel enabled:(BOOL)enabled;
 - (void) writeLEDThreshold:(int)channel;
 - (void) writeCFDParameters:(int)channel;
